+++ src/App/Bundle/MainBundle/Infrastructure/Submission/EntityConverter.php
@@ -24,6 +24,8 @@ class EntityConverter
             $entity->getGfycatKey(),
             $entity->getCreatedAt(),
             $entity->getBaseUrl(),
+            $entity->getWebmUrl(),
+            $entity->getMp4Url(),
             $entity->getAuthor()
         );
         $submission->setIdentifier($entity->getIdentifier());
@@ -45,6 +47,8 @@ class EntityConverter
             ->setGfycatKey($submission->getGfycatKey())
             ->setBaseUrl($submission->getBaseUrl())
             ->setCreatedAt($submission->getCreatedAt())
+            ->setWebmUrl($submission->getWebmUrl())
+            ->setMp4Url($submission->getMp4Url())
         ;
     }
 }