+++ src/App/Domain/Submission/Submission.php
@@ -45,25 +45,45 @@ class Submission
     /**
      * The author
      *
-     * @var string
+     * @var string|null
      */
     private $author;
 
     /**
+     * The webm URL
+     *
+     * @var string
+     */
+    private $webmUrl;
+
+    /**
+     * The mp4 URL
+     *
+     * @var string
+     */
+    private $mp4Url;
+
+    /**
      * __construct
      *
-     * @param string    $title
-     * @param string    $gfycatKey
-     * @param \DateTime $createdAt
-     * @param string    $baseUrl
-     * @param string    $author
+     * @param string      $title
+     * @param string      $gfycatKey
+     * @param \DateTime   $createdAt
+     * @param string      $baseUrl
+     * @param string      $webmUrl
+     * @param string      $mp4Url
+     * @param string|null $author
      */
-    public function __construct($title, $gfycatKey, \DateTime $createdAt, $baseUrl, $author)
-    {
+    public function __construct(
+        $title, $gfycatKey, \DateTime $createdAt,$baseUrl, $webmUrl, $mp4Url,
+        $author = null
+    ) {
         $this->title     = $title;
         $this->gfycatKey = $gfycatKey;
         $this->createdAt = $createdAt;
         $this->baseUrl   = $baseUrl;
+        $this->webmUrl   = $webmUrl;
+        $this->mp4Url    = $mp4Url;
         $this->author    = $author;
     }
 
@@ -122,7 +142,7 @@ class Submission
     /**
      * Get author
      *
-     * @return string
+     * @return string|null
      */
     public function getAuthor()
     {
@@ -137,5 +157,25 @@ class Submission
     public function getBaseUrl()
     {
         return $this->baseUrl;
+    }
+
+    /**
+     * Get webmUrl
+     *
+     * @return string
+     */
+    public function getWebmUrl()
+    {
+        return $this->webmUrl;
+    }
+
+    /**
+     * Get mp4Url
+     *
+     * @return string
+     */
+    public function getMp4Url()
+    {
+        return $this->mp4Url;
     }
 }